+++ CHANGELOG.md
@@ -5,6 +5,7 @@
         * Rename `zonedb` to `zonedball`. Consumes 72050 bytes of flash.
         * Repurpose `zonedb` to support only year 2000 and onwards.
           Consumes 34948 bytes of flash, less than 1/2 of `zonedball`.
+    * Replace out-of-band `isFilled` with in-band `year!=InvalidYear`.
 * 0.3.0 (2023-03-10, TZDB 2022g)
     * Convert C++ doxygen docs to go doc format.
     * Unexport various internal functions and structs by lowercasing their